+++ server/migrations/2021-06-03-063617_portfolios/up.sql
@@ -1,13 +1,13 @@
 CREATE TABLE IF NOT EXISTS portfolios (
-  id SERIAL NOT NULL PRIMARY KEY,
+  id BIGSERIAL NOT NULL PRIMARY KEY,
   slug VARCHAR NOT NULL,
   name VARCHAR,
-  owner INTEGER,
+  owner bigint,
   FOREIGN KEY (owner) REFERENCES identities (id) ON DELETE CASCADE
 );
 
 CREATE TABLE IF NOT EXISTS portfolio_assets (
-  portfolio SERIAL NOT NULL,
+  portfolio BIGSERIAL NOT NULL,
   asset VARCHAR NOT NULL,
   PRIMARY KEY (portfolio, asset),
   FOREIGN KEY (portfolio) REFERENCES portfolios (id) ON DELETE CASCADE,
