+++ server/src/asset/models/category.rs
@@ -3,12 +3,12 @@ use crate::db::schema::asset_categories;
 #[derive(Queryable, Clone, Debug)]
 pub struct Category {
     pub id: String,
-    pub owner: Option<i32>,
+    pub owner: Option<i64>,
 }
 
 #[derive(Insertable)]
 #[table_name = "asset_categories"]
 pub struct NewCategory<'a> {
     pub id: &'a str,
-    pub owner: Option<i32>,
+    pub owner: Option<i64>,
 }
