+++ server/src/service/models.rs
@@ -5,7 +5,7 @@ pub struct Service {
     pub id: String,
     pub name: String,
     pub url: Option<String>,
-    pub owner: Option<i32>,
+    pub owner: Option<i64>,
 }
 
 #[derive(Insertable)]
@@ -14,7 +14,7 @@ pub struct NewService<'a> {
     pub id: &'a str,
     pub name: &'a str,
     pub url: Option<&'a str>,
-    pub owner: Option<i32>,
+    pub owner: Option<i64>,
 }
 
 #[derive(Queryable, Clone, Debug)]
@@ -23,7 +23,7 @@ pub struct Wallet {
     pub name: String,
     pub url: Option<String>,
     pub service: Option<String>,
-    pub owner: Option<i32>,
+    pub owner: Option<i64>,
 }
 
 #[derive(Insertable)]
@@ -33,5 +33,5 @@ pub struct NewWallet<'a> {
     pub name: Option<&'a str>,
     pub url: Option<&'a str>,
     pub service: Option<&'a str>,
-    pub owner: Option<i32>,
+    pub owner: Option<i64>,
 }
